# Numba accelerated alpha filtrations