A ridiculously preliminary version of a numba-accelerated version of `alpha.py`.  The `numba` dependency has not yet been added to the installation requirements.

There are two families of changes (besides some minor code cleanup):
1. Changes due to `numba`'s stricter requirements on types. One of the effects of this is that some class methods have to become separate functions. Furthermore, the warnings in the circumcenter function currently break `numba` compilation so I commented them out.
2. Miscellaneous algorithmic changes. Chiefly, I removed some redundant conditionals and the `else` clause in the for loop over `sigma`s. I also comment out some code which was unused, e.g. the variable assignment currently in line 198.

The commented out code in lines 112-127 are an alternative "global" version of the Gabriel condition check which I believe can shave off some further runtime due to fewer calls to the circumcenter function.