@sauln 2020-11-04
Sweet! I'd love to see some numbers on how this improves the speed! I assume it's big time.