@ulupo 2020-11-04
Indeed there is a speedup.  Sorry I don't have time right now to give you a comprehensive benchmark test, but here are some very quick numbers from my MacBook:
- **2D**: for `X = 2 * tadasets.dsphere(n=100, d=1, r=1, noise=0.2)`, from 41.1 ms to 16.5 ms;
- **3D**: for `X = 2 * tadasets.dsphere(n=100, d=2, r=1, noise=0.2)`, from 7.68 s to 2.2 s or even 1.76 s when checking the Gabriel condition globally (i.e. using the commented-out alternative for cofaces).