@ulupo 2021-01-26
Update: the latest incarnation improves upon the version on `master` as follows on my machine:

**4D**
- 10000 pts: for X = 2 * tadasets.dsphere(n=10000, d=3, r=1, noise=0.2): from ~21min 5s to ~1min 50s.
- 1000 pts: for X = 2 * tadasets.dsphere(n=1000, d=3, r=1, noise=0.2): from ~1min 45s ± 137ms to 8.88s ± 105ms.
- 100 pts : for X = 2 * tadasets.dsphere(n=100, d=3, r=1, noise=0.2): from 6.18s ± 56.5ms to 523ms ± 11.6ms.
This is a **~11.8x** linear speedup. 

**3D**
- 10000 pts: for X = 2 * tadasets.dsphere(n=10000, d=2, r=1, noise=0.2): from ~1min 26s to ~11.6s.
- 1000 pts: for X = 2 * tadasets.dsphere(n=1000, d=2, r=1, noise=0.2): from 8.13s ± 137ms to 1.1s ± 16.6ms.
- 100 pts : for X = 2 * tadasets.dsphere(n=100, d=2, r=1, noise=0.2): from 660ms ± 21.2ms to 88.3ms ± 2.23ms.
This is a **~7.5x** linear speedup. 

**2D**
- 10000 pts: for X = 2 * tadasets.dsphere(n=10000, d=1, r=1, noise=0.2): from 4.71s ± 37.5ms to 1.18s ± 34ms.
- 1000 pts: for X = 2 * tadasets.dsphere(n=1000, d=1, r=1, noise=0.2): from 464ms ± 7.28ms to 114ms ± 3.13ms.
- 100 pts : for X = 2 * tadasets.dsphere(n=100, d=1, r=1, noise=0.2): from 44.9ms ± 2ms to 11.1ms ± 204µs.
This is a **~4x** linear speedup.

The extra performance comes *mainly* from the following changes:
1. New `numba`-accelerated code for the computation of the circumradius and/or the circumcentre, leveraging [Miroslav Fiedler's method](https://math.stackexchange.com/questions/2414640/circumsphere-of-a-tetrahedron/2415021#2415021) and using a jitted implementation of the pairwise distance function (considerably faster than `scipy` counterparts such as `pdist` + `squareform` or `cdist`, for small matrices).
2. Avoiding duplicate computations: all simplices in a non-top dimension `dim` can be stored as dictionary keys while traversing the set of simplices of dimension `dim + 1`.  Henceforth, they can be efficiently enumerated later on without repetition.  Currently, circumcircle calculations are duplicated because a simplex typically has several cofaces.

One notable missing feature is any sort of treatment of numerical issues coming from lack of general position.  This is still WIP with @ctralie.  And correctness even in the presence of general position still needs to be checked more thoroughly!

Another frontier is jitting not just the workhorse circumcircle functions but also the main algorithm.  There is a technical issue here in that `numba` does not currently accept variable-length tuples in `nopython` mode, but I might be able to find a workaround.