@ulupo 2021-01-28
I was finally able to njit everything!  As an example of the latest figures, I can now run 10000 points in 4 dimensions in about 25 s.  The main bottleneck is in fact a "stupid" list comprehension doing little more than dictionary unpacking and happening outside of the Numba code.  At the moment, this can take up to almost 50% of the whole execution time!